+++ app/src/main/java/blog/mazleo/ruvacant/core/ActivityModule.java
@@ -10,6 +10,7 @@ import androidx.fragment.app.FragmentManager;
 import blog.mazleo.ruvacant.R;
 import blog.mazleo.ruvacant.core.ApplicationAnnotations.ContentBodyFragment;
 import blog.mazleo.ruvacant.core.ApplicationAnnotations.ContentBodyParentFragment;
+import blog.mazleo.ruvacant.core.ApplicationAnnotations.ContentBodyTopBar;
 import blog.mazleo.ruvacant.ui.ApplicationActivity;
 import dagger.Module;
 import dagger.Provides;
@@ -63,5 +64,12 @@ public abstract class ActivityModule {
   @Nullable
   static View provideContentBodyParentFragment(ApplicationActivity activity) {
     return activity.findViewById(R.id.content_body_parent);
+  }
+
+  @Provides
+  @ContentBodyTopBar
+  @Nullable
+  static View provideContentBodyTopBar(ApplicationActivity activity) {
+    return activity.findViewById(R.id.content_body_top_bar);
   }
 }