+++ app/src/main/java/blog/mazleo/ruvacant/core/ApplicationAnnotations.java
@@ -25,5 +25,10 @@ public class ApplicationAnnotations {
   @Qualifier
   @Retention(RetentionPolicy.RUNTIME)
   @Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
+  public @interface ContentBodyTopBar {}
+
+  @Qualifier
+  @Retention(RetentionPolicy.RUNTIME)
+  @Target({ElementType.FIELD, ElementType.METHOD, ElementType.PARAMETER})
   public @interface MainThread {}
 }